Creating and using CustomerList Class instead of list
A new class was created which contains all customers. Before this change, a list was used for the customers. The new class called CustomerList is derived from list<T>. The reason for the change was because the id handling was put outside the class Customer. Now the id handling is done by CustomerList because an id handling in Customer class causes problems.